Split parametric_instance.rs
