+++ python/ommx/src/lib.rs
@@ -13,6 +13,8 @@ mod function;
 mod instance;
 mod linear;
 mod mps;
+mod parameters;
+mod parametric_instance;
 mod polynomial;
 mod qplib;
 mod quadratic;
@@ -39,6 +41,8 @@ pub use function::*;
 pub use instance::*;
 pub use linear::*;
 pub use mps::*;
+pub use parameters::*;
+pub use parametric_instance::*;
 pub use polynomial::*;
 pub use qplib::*;
 pub use quadratic::*;
